Remove useCircleMotion hook
